Enable Credhub within Cloud Foundry
This is a second Credhub, separate to the one used for our BOSH
manifests and Concourse pipelines. This second Credhub will be used
for tenant secrets.

This commit also sets up the rest of Cloud Foundry to use the new
Credhub to resolve `credhub-ref` entries in `VCAP_SERVICES`. Ready for
a Credhub Service Broker to be deployed, and other service brokers to
start working via Credhub.